back/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cohere==5.15.0
colorama==0.4.6
fastapi==0.115.12
fastavro==1.10.0
filelock==3.18.0
fsspec==2025.3.2
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
iniconfig==2.1.0
joblib==1.4.2
newsapi-python==0.2.7
nltk==3.9.1
numpy==2.2.5
packaging==25.0
pluggy==1.5.0
pydantic==2.11.3
pydantic_core==2.33.1
pytest==8.3.5
pytest-asyncio==0.26.0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
sniffio==1.3.1
starlette==0.46.2
threadpoolctl==3.6.0
tokenizers==0.21.1
tqdm==4.67.1
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
Unidecode==1.3.8
urllib3==2.4.0
uvicorn==0.34.1
